Add 1.18.2 support.

+++ 1_13_R1/src/main/java/de/rapha149/voidtotem/version/Wrapper1_13_R1.java
@@ -83,7 +83,7 @@ public class Wrapper1_13_R1 implements VersionWrapper {
     }
 
     @Override
-    public float getAbsorptionHearts(Player player) {
+    public double getAbsorptionHearts(Player player) {
         return ((CraftPlayer) player).getHandle().getAbsorptionHearts();
     }
 }
